Split common tests
