--- test/utils/common.go
@@ -1,11 +1,0 @@
-package utils
-
-import (
-	"testing"
-)
-
-func TestCommon(t *testing.T, params Params) {
-	TestConfig(t, params.Snap, params.Config)
-	TestNet(t, params.Snap, params.Net)
-	TestPackaging(t, params.Snap, params.Packaging)
-}
